+++ src/com/aussipvp/entity/projectile/Projectile.java
@@ -7,7 +7,7 @@ import com.aussipvp.graphics.Sprite;
 
 public abstract class Projectile extends Entity {
 
-	final protected int xOrigin, yOrigin;
+	final protected double xOrigin, yOrigin;
 	protected double angle;
 	protected Sprite sprite;
 	protected double x, y;
@@ -16,7 +16,7 @@ public abstract class Projectile extends Entity {
 	protected double speed, range, damage;
 	protected final Random random = new Random();
 
-	public Projectile(int x, int y, double dir) {
+	public Projectile(double x, double y, double dir) {
 		this.xOrigin = x;
 		this.yOrigin = y;
 		this.angle = dir;
